refactor(backend): split comparisons.py into sub-package
Decomposes the 861-LOC phenopackets/routers/comparisons.py into:

  routers/comparisons/
    __init__.py      (37)  re-exports router + stat helpers
    schemas.py       (55)  PhenotypeComparison, ComparisonResult
    statistics.py    (94)  Fisher's exact + BH FDR + Cohen's h
    variant_sql.py   (205) per-mode variant classification SQL fragments
    query.py         (120) phenotype-distribution CTE assembly
    router.py        (188) thin FastAPI /compare/variant-types endpoint

Every file is under 500 LOC. The statistical helpers are re-exported
at the package level so the 1,467-line tests/test_comparisons.py
keeps its existing imports unchanged:

    from app.phenopackets.routers.comparisons import (
        calculate_cohens_h,
        calculate_fdr_correction,
        calculate_fisher_exact_test,
    )

HTTP surface is byte-identical — locked in by the Wave 4 baseline
fixture phenopackets_compare_variant_types.json and by the existing
test_comparisons.py regression suite.

The biggest simplification is variant_sql.py: the original file had
the same ~90-line truncating-classification SQL duplicated across
three branches of an if/elif chain. build_group_conditions() now
returns it from a single _TRUNCATING_BASE constant, with the
large-CNV exclusion branch composing it via NOT + AND rather than
re-pasting.

Closes P2 #5 line item for comparisons.py.

Counts unchanged: 907 passed, 9 skipped, 3 xfailed.
